# Optimize netloc extraction