--- a/tests/main_test.py
+++ b/tests/main_test.py
@@ -205,6 +205,11 @@ def test_empty():
 
 
 def test_scheme():
+    assert_extract("//", ("", "", "", ""))
+    assert_extract("://", ("", "", "", ""))
+    assert_extract("://example.com", ("", "", "", ""))
+    assert_extract("a+-.://example.com", ("example.com", "", "example", "com"))
+    assert_extract("#//example.com", ("", "", "", ""))
     assert_extract(
         "https://mail.google.com/mail", ("mail.google.com", "mail", "google", "com")
     )